Jack's rearrangement - handing over to Luke. Moved script_gen into utils. Moved datasets to its own folder. Got rid of run_epoch.
