--- script_generation_tools/script_templates/cluster_multi_gpu_charles_template_script.sh
@@ -1,40 +1,0 @@
-#!/bin/sh
-#SBATCH -N 1	  # nodes requested
-#SBATCH -n 1	  # tasks requested
-#SBATCH --gres=gpu:2
-#SBATCH --mem=12000  # memory in Mb
-#SBATCH --time=3-00:00:00
-#SBATCH --exclude=charles[03-11]
-
-export CUDA_HOME=/opt/cuda-9.2.148/
-
-export CUDNN_HOME=/opt/cuDNN-7.0/
-
-export STUDENT_ID=$(whoami)
-
-export LD_LIBRARY_PATH=${CUDNN_HOME}/lib64:${CUDA_HOME}/lib64:$LD_LIBRARY_PATH
-
-export LIBRARY_PATH=${CUDNN_HOME}/lib64:$LIBRARY_PATH
-
-export CPATH=${CUDNN_HOME}/include:$CPATH
-
-export PATH=${CUDA_HOME}/bin:${PATH}
-
-export PYTHON_PATH=$PATH
-
-mkdir -p /disk/scratch/${STUDENT_ID}
-
-
-export TMPDIR=/disk/scratch/${STUDENT_ID}/
-export TMP=/disk/scratch/${STUDENT_ID}/
-
-mkdir -p ${TMP}/datasets/
-export DATASET_DIR=${TMP}/datasets/
-# Activate the relevant virtual environment:
-
-
-source /home/${STUDENT_ID}/miniconda3/bin/activate
-conda activate mlp
-
-cd ..
-python $execution_script$ --filepath_to_arguments_json_file experiment_config_files/$experiment_config$ -ngpus 2 -w 4
